MDEQ Beach Stations

Station information for MDEQ's recreational beach sampling program

Usage

MDEQ_beach_stations

Format

An sf object, using CRS NAD83 (EPSG:4269) with 21 features (rows) and 3 fields (columns), with point geometry for each.

DEQStationID

character, station ID matching that on DEQ's beach website. In WQX, this station is provided as part of the 'MonitoringLocationDescriptionText' field.

WQXStationID

character, station ID matching that in the 'MonitoringLocationIdentifier field from a Water Quality Portal download.

StationName

character, station name as converted to title case from all caps in the 'MonitoringLocationName' field of a Water Quality Portal download. This corresponds to the station names as given on the DEQ beach website, with some minor differences that do not affect station identification (e.g. 'Long Beach' here is 'Long Beach Beach' on the beach website).

geometry

POINT, degrees longitude and latitude, obtained from WQX.

Source

Combination of data download from a query of stations in the BCH24 project in the Water Quality Portal and cross-checking with stations on the MDEQ beach website.

Details

The Mississippi Department of Environmental Quality (MDEQ) regularly samples 21 stations across the MS coast for Enterococcus, a fecal indicator bacteria, as well as various nutrients, chlorophyll a, and Total Suspended Solids. Data is available through their beach website and through the Water Quality Portal, using Organization ID '21MSWQ_WQX' and Project ID 'BCHYY', where 'YY' is the last two digits of the year of interest.